feat: implementa exceções

+++ src/player/player.service.ts
@@ -1,13 +1,14 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreatePlayerDto } from './dto/create-player.dto';
 import { UpdatePlayerDto } from './dto/update-player.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { Player } from '@prisma/client';
 
 @Injectable()
 export class PlayerService {
   constructor(private prisma: PrismaService) {}
 
-  create(createPlayerDto: CreatePlayerDto) {
+  async create(createPlayerDto: CreatePlayerDto): Promise<Player> {
     return this.prisma.player.create({
       data: createPlayerDto,
       include: {
@@ -16,26 +17,45 @@ export class PlayerService {
     });
   }
 
-  findAll() {
-    return this.prisma.player.findMany({
+  async findAll(): Promise<Player[] | void> {
+    const players = await this.prisma.player.findMany({
       include: {
         team: true,
       },
     });
+
+    if (!players) {
+      throw new NotFoundException('Não existe nenhum jogador');
+    }
+    return players;
   }
 
-  findOne(id: number) {
-    return this.prisma.player.findMany({
+  async findOne(id: number): Promise<Player | void> {
+    const player = await this.prisma.player.findUnique({
       where: { id },
       include: {
-        team: {
-          select: { name: true },
-        },
+        team: true,
       },
     });
+
+    if (!player) {
+      throw new NotFoundException('Jogador não encontrado');
+    }
+    return player;
   }
 
-  update(id: number, updatePlayerDto: UpdatePlayerDto) {
+  async update(id: number, updatePlayerDto: UpdatePlayerDto): Promise<Player> {
+    const player = await this.prisma.player.findUnique({
+      where: { id },
+      include: {
+        team: true,
+      },
+    });
+
+    if (!player) {
+      throw new NotFoundException('Jogador não encontrado');
+    }
+
     return this.prisma.player.update({
       where: { id },
       data: updatePlayerDto,
@@ -45,7 +65,17 @@ export class PlayerService {
     });
   }
 
-  remove(id: number) {
+  async remove(id: number): Promise<Player> {
+    const player = await this.prisma.player.findUnique({
+      where: { id },
+      include: {
+        team: true,
+      },
+    });
+
+    if (!player) {
+      throw new NotFoundException('Jogador não encontrado');
+    }
     return this.prisma.player.delete({
       where: { id },
     });
