+++ src/team/team.service.ts
@@ -1,37 +1,66 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateTeamDto } from './dto/create-team.dto';
 import { UpdateTeamDto } from './dto/update-team.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { Team } from '@prisma/client';
 
 @Injectable()
 export class TeamService {
   constructor(private prisma: PrismaService) {}
-  create(createTeamDto: CreateTeamDto) {
+
+  async create(createTeamDto: CreateTeamDto): Promise<Team> {
     return this.prisma.team.create({
       data: createTeamDto,
     });
   }
 
-  findAll() {
-    return this.prisma.team.findMany();
+  async findAll(): Promise<Team[] | void> {
+    const teams = await this.prisma.team.findMany();
+    if (!teams) {
+      throw new NotFoundException('Não existe nenhum time');
+    }
+    return teams;
   }
 
-  findOne(id: number) {
-    return this.prisma.team.findUnique({
+  async findOne(id: number): Promise<Team | void> {
+    const team = await this.prisma.team.findUnique({
       where: { id },
     });
+
+    if (!team) {
+      throw new NotFoundException('Time não encontrado');
+    }
+    return team;
   }
 
-  update(id: number, updateTeamDto: UpdateTeamDto) {
+  async update(id: number, updateTeamDto: UpdateTeamDto): Promise<Team | void> {
+    const team = await this.prisma.team.findUnique({
+      where: { id },
+    });
+
+    if (!team) {
+      throw new NotFoundException('Time não encontrado');
+    }
     return this.prisma.team.update({
       where: { id },
       data: updateTeamDto,
     });
   }
 
-  remove(id: number) {
-    return this.prisma.team.delete({
+  async remove(id: number): Promise<Team> {
+    const team = await this.prisma.team.findUnique({
       where: { id },
     });
+
+    if (!team) {
+      throw new NotFoundException('Time não encontrado');
+    }
+    const deleteTeam = await this.prisma.team.delete({
+      where: { id },
+    });
+    if (!deleteTeam) {
+      throw new Error('Time esta relacionado a um jogador');
+    }
+    return deleteTeam;
   }
 }
